refactor(pkg): make public pkg internal
