+++ cmd/completion.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/spf13/cobra"
 
-	"github.com/alex-held/devctl/pkg/cli"
+	"github.com/alex-held/devctl/internal/cli"
 )
 
 // NewCompletionCommand Creates new Completion cobra.Command
