+++ internal/cli/options_test.go
